conquer_project/requirements1.txt
# Backend
fastapi
uvicorn[standard]
scikit-learn
pandas
nltk
faiss-cpu
transformers
torch
python-multipart

# Dashboard
streamlit
requests